Moving connectionString and communicationClientCredential into impl folder
